entropy: remove m_slice member from SBac class
The slice object was only needed in resetEntropy() and determineCabacInitIdx()
and could be passed to those two methods easily.

It was also being used in two functions to gain access to the scaling list, but
this was actually a problem for us because headers can be queuried before any
slice objects are initialized so it was better to pass in the scaling lists
directly from the top Encoder. Note this fixes a bug in scaling list support -
we were previously incapable of signaling them properly in our stream headers.

Removing this member variable lets us clean up a lot of redundant setSlice()
calls and other cruft