rc: update ratecontrol stats in every frame, avoid frame parallelism lag in abr
RateControl statistics are updated for every frame when refLagRows number of
rows are completed in processRowEncoder. With this updated data rateControl
predicts more accurate QP

This removes the previous hack which disabled frame parallelism for the first
half-second of video and replaces it with a more fine-grained rate control
call timing to achieve a similar ABR improvement with less effect to frame
parallelism.