quant: change how RDOQ measures distortion [CHANGES OUTPUTS]
RDOQ, as it was written in the HM, expects scaled level values to be output by
quant; these are the output levels multiplied by the quantizing coeffificient
but without the rounding factor and without the downshift. It would then
measure distortion as the difference between this scaled level and level <<
qbits (a rough unquant). To make this math work, it was pre-calculating an
error scale factor (per block position, since the quantization coefficients can
vary) which divided the result by the squared scale factor and upshifting to
simultaneously account for the FIX15 nature of the signaling costs and the
uniform scaling of the forward transform. To roughly summarize:

   errScale = (1 << (15 - 2 * transformShift)) / (quantCoeff[i] * quantCoeff[i])
   levelScaled = level * quantCoeff[i]
   distortion = levelScaled - (level << qbits);
   cost = distortion * distortion * errScale + lambda2 * bitsFix15

It was forced to use floating point math for the errScale and distortion
calculations, and thus did not bother with fixed point math for lambda2.

This commit changes the distortion measurement to be the difference between the
original (pre-quantization) DCT coefficient and the unquantized level.

   unquantAbsLevel = (level * quantCoeff[i] + pad) >> shift;
   distortion = unquantAbsLevel - abs(signCoef);
   distScale = 1 << (15 - 2 * transformShift);
   cost = distortion * distortion << distScale + lambda2 * bitsFix15

Note that the same scale factor is still required to account for the FIX15 bit
cost and the forward DCT scale but now it is a simple shift operation.

This commit does not change the data types; that will be a later commit once the
dynamic ranges have been properly evaluated.  And deltaU[], used by sign hiding,
is left using the scaled level cost basis for now.