frameencoder: avoid race hazard in end of frame logic
This bug has been in the encoder, latent, for about 6 months so it deserves an
obituary (being somewhat presumptuous that this commit actually kills it). The
bug worked in this way:

1. user starts an encode of a small resolution encode at superfast preset,
   generating frame rates of about 200-300 fps depending on the hardware.

2. a worker thread finishes compressing all of the CTUs in its row, but then
   gets evicted from its core before it is allowed to flush its row bitstream
   object (leaving four bytes within the Entropy object unwritten)

3. The other threads quickly finish the rest of the frame and the worker which
   compressed the last row signals frame completion.

4. The frame encoder thread concatenates the row bitstreams together and emits
   the NAL. If you were lucky, the row that was evicted would have been 4 bytes
   long since the 0 byte length triggers a check failure indicating something is
   seriously wrong. If you're un-lucky the NAL is garbage and does evil things
   to your decoder.

5. the evicted worker finally gets back on the core and flushes its bitstream,
   to the aid of no-one.

With this commit, we note when the last row has been compressed (indicating all
CTUs in the frame are compressed) but we do not signal that the frame is
complete until the last worker thread leaves processRow() (taking advantage of
the recently introduced atomic count of workers).